requirements.txt
json2html==1.3.0
numpy==1.24.4
#opencv_contrib_python_headless==4.8.0.74
#opencv_python_headless==4.8.0.74
opencv_contrib_python==4.8.1.78
opencv_python==4.8.1.78
openpyxl==3.1.2
python-dateutil==2.8.2
pandas==2.0.3
plotly==5.15.0
pyyaml==6.0.1
rospkg==1.5.0
defusedxml==0.7.1
gpiozero==2.0
gphoto2==2.5.0